requirements.txt
Flask==2.3.1
googletrans==4.0.0rc1
iso639-lang==2.1.0
PyYAML==6.0.1
replit==3.3.1